+++ tests/run-emitter.c
@@ -262,7 +262,7 @@ main(int argc, char *argv[])
         int done = 0;
         int count = 0;
         int error = 0;
-        int k;
+        size_t k;
         memset(buffer, 0, BUFFER_SIZE+1);
         memset(events, 0, MAX_EVENTS*sizeof(yaml_event_t));
 
@@ -319,7 +319,7 @@ main(int argc, char *argv[])
             yaml_parser_delete(&parser);
         }
 
-        for (k = 0; k < (int)event_number; k ++) {
+        for (k = 0; k < event_number; k ++) {
             yaml_event_delete(events+k);
         }
